requirements.txt
customtkinter~=5.0.3
PyAutoGUI~=0.9.53
PyGetWindow~=0.0.9
Pillow~=8.4.0
numpy~=1.19.5
tensorflow~=2.6.2
keras~=2.6.0
pandas~=1.1.5
matplotlib~=3.3.4
scikit-learn~=0.24.2
colorama~=0.4.5
